backend/requirements.txt
fastapi
uvicorn
torch
torchvision
timm
opencv-python
pillow
numpy